requirements.txt
requests>=2.0,<3.0
python_dateutil>=2.5.3
PyJWT >=1.7.1